+++ src/main/kotlin/com/atlassian/performance/tools/aws/api/ami/AmiProvider.kt
@@ -3,7 +3,7 @@ package com.atlassian.performance.tools.aws.api.ami
 import com.atlassian.performance.tools.aws.api.Aws
 
 /**
- * @since v1.10.1
+ * @since 1.10.1
  */
 interface AmiProvider {