+++ src/main/kotlin/com/atlassian/performance/tools/aws/api/ami/NoAmiCache.kt
@@ -5,7 +5,7 @@ import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 
 /**
- * @since v1.10.1
+ * @since 1.10.1
  */
 class NoAmiCache : SshAmiMod.AmiCache {